Add new and default constructors to Customer and Product
Added a constructor to the Customer class that initializes Email
and Password from a LoginViewModel object. Also added a default
constructor to the Customer class. Similarly, added a constructor
to the Product class that initializes Name, Description, Price,
Quantity, and Availability from a ProductViewModel object, along
with a default constructor for the Product class.